+++ src/Command/MappingNode/ListMappingNodeSiblings.php
@@ -5,9 +5,13 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\MappingNode;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\StringCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Overview\IdentityOverviewCriteria;
+use Heptacom\HeptaConnect\Storage\Base\Action\IdentityRedirect\Overview\IdentityRedirectOverviewCriteria;
+use Heptacom\HeptaConnect\Storage\Base\Action\IdentityRedirect\Overview\IdentityRedirectOverviewResult;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverviewActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -23,13 +27,17 @@ class ListMappingNodeSiblings extends Command
 
     private IdentityOverviewActionInterface $identityOverviewAction;
 
+    private IdentityRedirectOverviewActionInterface $identityRedirectOverviewAction;
+
     public function __construct(
         StorageKeyGeneratorContract $storageKeyGenerator,
-        IdentityOverviewActionInterface $identityOverviewAction
+        IdentityOverviewActionInterface $identityOverviewAction,
+        IdentityRedirectOverviewActionInterface $identityRedirectOverviewAction
     ) {
         parent::__construct();
         $this->storageKeyGenerator = $storageKeyGenerator;
         $this->identityOverviewAction = $identityOverviewAction;
+        $this->identityRedirectOverviewAction = $identityRedirectOverviewAction;
     }
 
     protected function configure(): void
@@ -45,12 +53,18 @@ class ListMappingNodeSiblings extends Command
         $entityType = (string) $input->getOption('entity-type');
         $portalNodeKeyParam = (string) $input->getOption('portal-node-key');
         $externalIds = (array) $input->getArgument('external-ids');
-        $criteria = new IdentityOverviewCriteria();
+        $identityCriteria = new IdentityOverviewCriteria();
+        $sourceIdentityRedirectCriteria = new IdentityRedirectOverviewCriteria();
 
-        if ($entityType !== '' && !\is_a($entityType, DatasetEntityContract::class, true)) {
-            $io->error('The provided dataset entity class does not implement the DatasetEntityContract.');
+        if ($entityType !== '') {
+            if (!\is_a($entityType, DatasetEntityContract::class, true)) {
+                $io->error('The provided dataset entity class does not implement the DatasetEntityContract.');
 
-            return 1;
+                return 1;
+            }
+
+            $identityCriteria->setEntityTypeFilter([$entityType]);
+            $sourceIdentityRedirectCriteria->setEntityTypeFilter(new StringCollection($identityCriteria->getEntityTypeFilter()));
         }
 
         if ($portalNodeKeyParam !== '') {
@@ -62,7 +76,8 @@ class ListMappingNodeSiblings extends Command
                 return 2;
             }
 
-            $criteria->getPortalNodeKeyFilter()->push([$portalNodeKey]);
+            $identityCriteria->getPortalNodeKeyFilter()->push([$portalNodeKey]);
+            $sourceIdentityRedirectCriteria->setSourcePortalNodeKeyFilter($identityCriteria->getPortalNodeKeyFilter());
         }
 
         $externalIds = \array_filter($externalIds);
@@ -73,7 +88,10 @@ class ListMappingNodeSiblings extends Command
             return 3;
         }
 
-        $criteria->setExternalIdFilter($externalIds);
+        $identityCriteria->setExternalIdFilter($externalIds);
+        $sourceIdentityRedirectCriteria->setSourceExternalIdFilter(new StringCollection($identityCriteria->getExternalIdFilter()));
+
+        $rows = [];
 
         $othersCriteria = new IdentityOverviewCriteria();
         $othersCriteria->setSort([
@@ -81,19 +99,74 @@ class ListMappingNodeSiblings extends Command
             IdentityOverviewCriteria::FIELD_MAPPING_NODE => IdentityOverviewCriteria::SORT_ASC,
             IdentityOverviewCriteria::FIELD_PORTAL_NODE => IdentityOverviewCriteria::SORT_ASC,
         ]);
+        $sourceIdentityRedirectCriteria->setSort([
+            IdentityRedirectOverviewCriteria::FIELD_ENTITY_TYPE => IdentityRedirectOverviewCriteria::SORT_ASC,
+            IdentityRedirectOverviewCriteria::FIELD_TARGET_PORTAL_NODE => IdentityRedirectOverviewCriteria::SORT_ASC,
+            IdentityRedirectOverviewCriteria::FIELD_TARGET_EXTERNAL_ID => IdentityRedirectOverviewCriteria::SORT_ASC,
+        ]);
 
-        foreach ($this->identityOverviewAction->overview($criteria) as $identity) {
+        foreach ($this->identityOverviewAction->overview($identityCriteria) as $identity) {
             $othersCriteria->getMappingNodeKeyFilter()->push([$identity->getMappingNodeKey()]);
         }
 
-        $rows = [];
+        if (!$othersCriteria->getMappingNodeKeyFilter()->isEmpty()) {
+            foreach ($this->identityOverviewAction->overview($othersCriteria) as $identity) {
+                $rows[] = [
+                    'portal-node-key' => $this->storageKeyGenerator->serialize($identity->getPortalNodeKey()->withAlias()),
+                    'external-id' => $identity->getExternalId(),
+                    'group-key' => $this->storageKeyGenerator->serialize($identity->getMappingNodeKey()),
+                    'entity-type' => $identity->getEntityType(),
+                ];
+            }
+        }
 
-        foreach ($this->identityOverviewAction->overview($othersCriteria) as $identity) {
+        $groupKeys = [];
+
+        /** @var IdentityRedirectOverviewResult $identityRedirect */
+        foreach ($this->identityRedirectOverviewAction->overview($sourceIdentityRedirectCriteria) as $identityRedirect) {
+            $groupKey = $this->storageKeyGenerator->serialize($identityRedirect->getIdentityRedirectKey());
+
             $rows[] = [
-                'portal-node-key' => $this->storageKeyGenerator->serialize($identity->getPortalNodeKey()->withAlias()),
-                'external-id' => $identity->getExternalId(),
-                'mapping-node-key' => $this->storageKeyGenerator->serialize($identity->getMappingNodeKey()),
-                'entity-type' => $identity->getEntityType(),
+                'portal-node-key' => $this->storageKeyGenerator->serialize($identityRedirect->getSourcePortalNodeKey()->withAlias()),
+                'external-id' => $identityRedirect->getSourceExternalId(),
+                'group-key' => $groupKey,
+                'entity-type' => $identityRedirect->getEntityType(),
+            ];
+            $rows[] = [
+                'portal-node-key' => $this->storageKeyGenerator->serialize($identityRedirect->getTargetPortalNodeKey()->withAlias()),
+                'external-id' => $identityRedirect->getTargetExternalId(),
+                'group-key' => $groupKey,
+                'entity-type' => $identityRedirect->getEntityType(),
+            ];
+
+            $groupKeys[] = $groupKey;
+        }
+
+        $targetIdentityRedirectCriteria = new IdentityRedirectOverviewCriteria();
+        $targetIdentityRedirectCriteria->setSort($sourceIdentityRedirectCriteria->getSort());
+        $targetIdentityRedirectCriteria->setTargetExternalIdFilter($sourceIdentityRedirectCriteria->getSourceExternalIdFilter());
+        $targetIdentityRedirectCriteria->setTargetPortalNodeKeyFilter($sourceIdentityRedirectCriteria->getSourcePortalNodeKeyFilter());
+        $targetIdentityRedirectCriteria->setEntityTypeFilter($sourceIdentityRedirectCriteria->getEntityTypeFilter());
+
+        /** @var IdentityRedirectOverviewResult $identityRedirect */
+        foreach ($this->identityRedirectOverviewAction->overview($targetIdentityRedirectCriteria) as $identityRedirect) {
+            $groupKey = $this->storageKeyGenerator->serialize($identityRedirect->getIdentityRedirectKey());
+
+            if (\in_array($groupKey, $groupKeys, true)) {
+                continue;
+            }
+
+            $rows[] = [
+                'portal-node-key' => $this->storageKeyGenerator->serialize($identityRedirect->getSourcePortalNodeKey()->withAlias()),
+                'external-id' => $identityRedirect->getSourceExternalId(),
+                'group-key' => $groupKey,
+                'entity-type' => $identityRedirect->getEntityType(),
+            ];
+            $rows[] = [
+                'portal-node-key' => $this->storageKeyGenerator->serialize($identityRedirect->getTargetPortalNodeKey()->withAlias()),
+                'external-id' => $identityRedirect->getTargetExternalId(),
+                'group-key' => $groupKey,
+                'entity-type' => $identityRedirect->getEntityType(),
             ];
         }